pyproject.toml
[project]
name = "pixelpath"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "flask>=3.1.1",
    "imageio>=2.37.0",
    "numpy>=2.2.6",
    "ordered-set>=4.1.0",
    "pillow>=11.2.1",
    "werkzeug>=3.1.3",
]
